Unification of naming convention of groupId for 1.0.x #150
