+++ .travis.yml
@@ -14,17 +14,17 @@ env:
     - CARGO_DAEMON_WEBAPP_VERSION=1.4.17
   matrix:
     # todo
-    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-archetype GROUP_ID=todo ARTIFACT_ID=todo
-    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-mybatis2-archetype GROUP_ID=todo ARTIFACT_ID=todo-mybatis2
-    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-mybatis2-archetype GROUP_ID=todo ARTIFACT_ID=todo-mybatis2-multi
-    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-jpa-archetype GROUP_ID=todo ARTIFACT_ID=todo-jpa
-    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-jpa-archetype GROUP_ID=todo ARTIFACT_ID=todo-jpa-multi
+    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo
+    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-mybatis2-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-mybatis2
+    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-mybatis2-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-mybatis2-multi
+    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-jpa-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-jpa
+    - TARGET_TUTORIAL=todo ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-jpa-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-jpa-multi
     # todo-rest
-    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-archetype GROUP_ID=todo ARTIFACT_ID=todo-api
-    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-mybatis2-archetype GROUP_ID=todo ARTIFACT_ID=todo-api-mybatis2
-    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-mybatis2-archetype GROUP_ID=todo ARTIFACT_ID=todo-api-mybatis2-multi
-    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-jpa-archetype GROUP_ID=todo ARTIFACT_ID=todo-api-jpa
-    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-jpa-archetype GROUP_ID=todo ARTIFACT_ID=todo-api-jpa-multi
+    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-api
+    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-mybatis2-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-api-mybatis2
+    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-mybatis2-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-api-mybatis2-multi
+    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-jpa-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-api-jpa
+    - TARGET_TUTORIAL=todo-rest ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-jpa-archetype GROUP_ID=com.example.todo ARTIFACT_ID=todo-api-jpa-multi
     # first-spring-security
     - TARGET_TUTORIAL=first-spring-security ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-web-blank-mybatis2-archetype GROUP_ID=com.example.security ARTIFACT_ID=first-spring-security-mybatis2
     - TARGET_TUTORIAL=first-spring-security ARCHETYPE_ARTIFACT_ID=terasoluna-gfw-multi-web-blank-mybatis2-archetype GROUP_ID=com.example.security ARTIFACT_ID=first-spring-security-mybatis2-multi
