--- todo-rest/src/main/java/todo/api/common/error/ApiError.java
@@ -1,66 +1,0 @@
-/*
- * Copyright (C) 2013-2017 NTT DATA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package todo.api.common.error;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.codehaus.jackson.map.annotate.JsonSerialize;
-import org.codehaus.jackson.map.annotate.JsonSerialize.Inclusion;
-
-public class ApiError {
-
-    private final String code;
-
-    private final String message;
-
-    @JsonSerialize(include = Inclusion.NON_EMPTY)
-    private final String target;
-
-    @JsonSerialize(include = Inclusion.NON_EMPTY)
-    private final List<ApiError> details = new ArrayList<>();
-
-    public ApiError(String code, String message) {
-        this(code, message, null);
-    }
-
-    public ApiError(String code, String message, String target) {
-        this.code = code;
-        this.message = message;
-        this.target = target;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public String getTarget() {
-        return target;
-    }
-
-    public List<ApiError> getDetails() {
-        return details;
-    }
-
-    public void addDetail(ApiError detail) {
-        details.add(detail);
-    }
-
-}
