--- todo-rest/src/main/java/todo/api/common/error/RestGlobalExceptionHandler.java
@@ -1,114 +1,0 @@
-/*
- * Copyright (C) 2013-2017 NTT DATA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package todo.api.common.error;
-
-import javax.inject.Inject;
-
-import org.springframework.context.MessageSource;
-import org.springframework.context.support.DefaultMessageSourceResolvable;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.validation.ObjectError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import org.terasoluna.gfw.common.exception.BusinessException;
-import org.terasoluna.gfw.common.exception.ResourceNotFoundException;
-import org.terasoluna.gfw.common.exception.ResultMessagesNotificationException;
-import org.terasoluna.gfw.common.message.ResultMessage;
-
-@ControllerAdvice
-public class RestGlobalExceptionHandler extends ResponseEntityExceptionHandler {
-
-    @Inject
-    MessageSource messageSource;
-
-    @Override
-    protected ResponseEntity<Object> handleExceptionInternal(Exception ex,
-            Object body, HttpHeaders headers, HttpStatus status,
-            WebRequest request) {
-        Object responseBody = body;
-        if (body == null) {
-            responseBody = createApiError(request, "E999", ex.getMessage());
-        }
-        return new ResponseEntity<Object>(responseBody, headers, status);
-    }
-
-    private ApiError createApiError(WebRequest request, String errorCode,
-            Object... args) {
-        return new ApiError(errorCode, messageSource.getMessage(errorCode,
-                args, request.getLocale()));
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleMethodArgumentNotValid(
-            MethodArgumentNotValidException ex, HttpHeaders headers,
-            HttpStatus status, WebRequest request) {
-        ApiError apiError = createApiError(request, "E400");
-        for (FieldError fieldError : ex.getBindingResult().getFieldErrors()) {
-            apiError.addDetail(createApiError(request, fieldError, fieldError
-                    .getField()));
-        }
-        for (ObjectError objectError : ex.getBindingResult().getGlobalErrors()) {
-            apiError.addDetail(createApiError(request, objectError, objectError
-                    .getObjectName()));
-        }
-        return handleExceptionInternal(ex, apiError, headers, status, request);
-    }
-
-    private ApiError createApiError(WebRequest request,
-            DefaultMessageSourceResolvable messageSourceResolvable,
-            String target) {
-        return new ApiError(messageSourceResolvable.getCode(), messageSource
-                .getMessage(messageSourceResolvable, request.getLocale()), target);
-    }
-
-    @ExceptionHandler(BusinessException.class)
-    public ResponseEntity<Object> handleBusinessException(BusinessException ex,
-            WebRequest request) {
-        return handleResultMessagesNotificationException(ex, null,
-                HttpStatus.CONFLICT, request);
-    }
-
-    private ResponseEntity<Object> handleResultMessagesNotificationException(
-            ResultMessagesNotificationException ex, HttpHeaders headers,
-            HttpStatus status, WebRequest request) {
-        ResultMessage message = ex.getResultMessages().iterator().next();
-        ApiError apiError = createApiError(request, message.getCode(), message
-                .getArgs());
-        return handleExceptionInternal(ex, apiError, headers, status, request);
-    }
-
-    @ExceptionHandler(ResourceNotFoundException.class)
-    public ResponseEntity<Object> handleResourceNotFoundException(
-            ResourceNotFoundException ex, WebRequest request) {
-        return handleResultMessagesNotificationException(ex, null,
-                HttpStatus.NOT_FOUND, request);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<Object> handleSystemError(Exception ex,
-            WebRequest request) {
-        ApiError apiError = createApiError(request, "E500");
-        return handleExceptionInternal(ex, apiError, null,
-                HttpStatus.INTERNAL_SERVER_ERROR, request);
-    }
-
-}
