--- todo-rest/src/main/java/todo/api/todo/TodoResource.java
@@ -1,66 +1,0 @@
-/*
- * Copyright (C) 2013-2017 NTT DATA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package todo.api.todo;
-
-import java.util.Date;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
-
-public class TodoResource {
-
-    private String todoId;
-
-    @NotNull
-    @Size(min = 1, max = 30)
-    private String todoTitle;
-
-    private boolean finished;
-
-    private Date createdAt;
-
-    public String getTodoId() {
-        return todoId;
-    }
-
-    public void setTodoId(String todoId) {
-        this.todoId = todoId;
-    }
-
-    public String getTodoTitle() {
-        return todoTitle;
-    }
-
-    public void setTodoTitle(String todoTitle) {
-        this.todoTitle = todoTitle;
-    }
-
-    public boolean isFinished() {
-        return finished;
-    }
-
-    public void setFinished(boolean finished) {
-        this.finished = finished;
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-}
