--- todo-rest/src/main/java/todo/api/todo/TodoRestController.java
@@ -1,92 +1,0 @@
-/*
- * Copyright (C) 2013-2017 NTT DATA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package todo.api.todo;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import javax.inject.Inject;
-
-import org.dozer.Mapper;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Controller;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import todo.domain.model.Todo;
-import todo.domain.service.todo.TodoService;
-
-@Controller
-@RequestMapping("todos")
-public class TodoRestController {
-    @Inject
-    TodoService todoService;
-    @Inject
-    Mapper beanMapper;
-
-    @RequestMapping(method = RequestMethod.GET)
-    @ResponseBody
-    @ResponseStatus(HttpStatus.OK)
-    public List<TodoResource> getTodos() {
-        Collection<Todo> todos = todoService.findAll();
-        List<TodoResource> todoResources = new ArrayList<>();
-        for (Todo todo : todos) {
-            todoResources.add(beanMapper.map(todo, TodoResource.class));
-        }
-        return todoResources;
-    }
-
-    @RequestMapping(method = RequestMethod.POST)
-    @ResponseBody
-    @ResponseStatus(HttpStatus.CREATED)
-    public TodoResource postTodos(@RequestBody @Validated TodoResource todoResource) {
-        Todo createdTodo = todoService.create(beanMapper.map(todoResource, Todo.class));
-        TodoResource createdTodoResponse = beanMapper.map(createdTodo, TodoResource.class);
-        return createdTodoResponse;
-    }
-
-    @RequestMapping(value="{todoId}", method = RequestMethod.GET)
-    @ResponseBody
-    @ResponseStatus(HttpStatus.OK)
-    public TodoResource getTodo(@PathVariable("todoId") String todoId) {
-        Todo todo = todoService.findOne(todoId);
-        TodoResource todoResource = beanMapper.map(todo, TodoResource.class);
-        return todoResource;
-    }
-
-    @RequestMapping(value="{todoId}", method = RequestMethod.PUT)
-    @ResponseBody
-    @ResponseStatus(HttpStatus.OK)
-    public TodoResource putTodo(@PathVariable("todoId") String todoId) {
-        Todo finishedTodo = todoService.finish(todoId);
-        TodoResource finishedTodoResource = beanMapper.map(finishedTodo, TodoResource.class);
-        return finishedTodoResource;
-    }
-
-    @RequestMapping(value="{todoId}", method = RequestMethod.DELETE) // (1)
-    @ResponseBody
-    @ResponseStatus(HttpStatus.NO_CONTENT) // (2)
-    public void deleteTodo(@PathVariable("todoId") String todoId) { // (3)
-        todoService.delete(todoId); // (4)
-    }
-
-}
