+++ todo/scripts/copy-sources.sh
@@ -10,24 +10,24 @@ pushd "${SCRIPT_DIR}/../target-project"
 if test `echo $ARCHETYPE_ARTIFACT_ID | grep multi`;then
   # multi project
   # web project
-  cp -p -r ../src/main/java/todo/app ./${ARTIFACT_ID}/${ARTIFACT_ID}-web/src/main/java/todo
+  cp -p -r ../src/main/java/com/example/todo/app ./${ARTIFACT_ID}/${ARTIFACT_ID}-web/src/main/java/com/example/todo
   cp -p -r ../src/main/webapp ./${ARTIFACT_ID}/${ARTIFACT_ID}-web/src/main
   # domain project
-  cp -p -r ../src/main/java/todo/domain ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/java/todo
+  cp -p -r ../src/main/java/com/example/todo/domain ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/java/com/example/todo
   # selenium project
   cp -p -r ../src/test ./${ARTIFACT_ID}/${ARTIFACT_ID}-selenium/src
   cp -p -r ../../common/src/test ./${ARTIFACT_ID}/${ARTIFACT_ID}-selenium/src
   
   case "${ARCHETYPE_ARTIFACT_ID}" in
     *mybatis2* )
-      cp -p -r ../src-mybatis2/main/java/todo/domain ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/java/todo
+      cp -p -r ../src-mybatis2/main/java/com/example/todo/domain ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/java/com/example/todo
       cp -p -r ../src-mybatis2/main/resources/META-INF/mybatis ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/resources/META-INF
       ;;
     *jpa* )
-      cp -p -r ../src-jpa/main/java/todo/domain ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/java/todo
+      cp -p -r ../src-jpa/main/java/com/example/todo/domain ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/java/com/example/todo
       ;;
     * )
-      cp -p -r ../src-plain/main/java/todo/domain ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/java/todo
+      cp -p -r ../src-plain/main/java/com/example/todo/domain ./${ARTIFACT_ID}/${ARTIFACT_ID}-domain/src/main/java/com/example/todo
       ;;
   esac
 else
