--- todo/src-jpa/main/java/todo/domain/model/Todo.java
@@ -1,82 +1,0 @@
-/*
- * Copyright (C) 2013-2017 NTT DATA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package todo.domain.model;
-
-import java.io.Serializable;
-import java.util.Date;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-// (1)
-@Entity
-@Table(name = "todo")
-public class Todo implements Serializable {
-    private static final long serialVersionUID = 1L;
-
-    // (2)
-    @Id
-    // (3)
-    @Column(name = "todo_id")
-    private String todoId;
-
-    @Column(name = "todo_title")
-    private String todoTitle;
-
-    @Column(name = "finished")
-    private boolean finished;
-
-    @Column(name = "created_at")
-    // (4)
-    @Temporal(TemporalType.TIMESTAMP)
-    private Date createdAt;
-
-    public String getTodoId() {
-        return todoId;
-    }
-
-    public void setTodoId(String todoId) {
-        this.todoId = todoId;
-    }
-
-    public String getTodoTitle() {
-        return todoTitle;
-    }
-
-    public void setTodoTitle(String todoTitle) {
-        this.todoTitle = todoTitle;
-    }
-
-    public boolean isFinished() {
-        return finished;
-    }
-
-    public void setFinished(boolean finished) {
-        this.finished = finished;
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-}
