--- todo/src-plain/main/java/todo/domain/repository/todo/TodoRepositoryImpl.java
@@ -1,60 +1,0 @@
-/*
- * Copyright (C) 2013-2017 NTT DATA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package todo.domain.repository.todo;
-
-import java.util.Collection;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import org.springframework.stereotype.Repository;
-
-import todo.domain.model.Todo;
-
-@Repository // (1)
-public class TodoRepositoryImpl implements TodoRepository {
-    private static final Map<String, Todo> TODO_MAP = new ConcurrentHashMap<String, Todo>();
-
-    @Override
-    public Todo findOne(String todoId) {
-        return TODO_MAP.get(todoId);
-    }
-
-    @Override
-    public Collection<Todo> findAll() {
-        return TODO_MAP.values();
-    }
-
-    @Override
-    public Todo save(Todo todo) {
-        return TODO_MAP.put(todo.getTodoId(), todo);
-    }
-
-    @Override
-    public void delete(Todo todo) {
-        TODO_MAP.remove(todo.getTodoId());
-    }
-
-    @Override
-    public long countByFinished(boolean finished) {
-        long count = 0;
-        for (Todo todo : TODO_MAP.values()) {
-            if (finished == todo.isFinished()) {
-                count++;
-            }
-        }
-        return count;
-    }
-}
