--- todo/src/main/java/todo/domain/service/todo/TodoService.java
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) 2013-2017 NTT DATA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package todo.domain.service.todo;
-
-import java.util.Collection;
-
-import todo.domain.model.Todo;
-
-public interface TodoService {
-    Collection<Todo> findAll();
-
-    Todo create(Todo todo);
-
-    Todo finish(String todoId);
-
-    void delete(String todoId);
-}
